requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
djangorestframework
django-cors-headers
django-extensions
